Add C code generator for points to plot
The python script uses SymPy to generate equations for position vectors
of points that are supposed to be plotted. The python script generates
three C codes, each should return a 12x2 array that corresponds to the
starting and ending points of that lines that are supposed to be drawn.
The CMakeLists.txt is updated to run this script before build starts.