Update the plot on the arrival of new data
Extremely bad performance compared to what was originally thought.
The main problem is the use of GnuPlot by the Matplot++. This causes
the figure to flicker a lot as the data for graph objects is changed.
Matplot++ does have an experimental OpenGL backend. However, it is
not super functional rightnow. It would't even generate a 3d line
graph. So, a better solution is definetly needed in the near future.